+++ view/frontend/requirejs-config.js
@@ -1,7 +1,4 @@
 var config = {
-    "paths": {
-        "webp-detect": "Yireo_Webp2/js/detect"
-    },
     config: {
         mixins: {
             'Magento_Swatches/js/swatch-renderer': {